feat(icon renderer): on dispose, schedule container gameobject to be destroyed instead of asking for it to be destroyed immediately
Originally this was also calling Destroy(), but I changed it to DestroyImmediate when debugging.
It can go back to just being a Destroy. I don't think Dispose() has the urgency to immediately get rid of stuff, and it aligns with other calls to GameObject.Destroy() in this and the text renderer's Dispose().